Changes to allow compilation on Linux and eliminate some coredumps.
In XbText, the size of the str buffer in XmPrintf  has been increased to accomodate longer chains.

In PropertiesTab, the dimension of _formatting_buffer2  was increased to permit printing
addresses > 32bits and avoiding coredumps.

In AccessTab corrected in AccessTabActive the display of hosts allowed to connect
to the X server.